Creating only one drove client per upstream
