+++ ranger-drove/src/main/java/io/appform/ranger/drove/common/DroveNodeDataStoreConnector.java
@@ -17,29 +17,10 @@ package io.appform.ranger.drove.common;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.phonepe.drove.client.DroveClient;
-import com.phonepe.drove.client.DroveClientConfig;
-import com.phonepe.drove.client.decorators.AuthHeaderDecorator;
-import com.phonepe.drove.client.decorators.BasicAuthDecorator;
-import com.phonepe.drove.client.transport.httpcomponent.DroveHttpComponentsTransport;
 import io.appform.ranger.core.model.NodeDataStoreConnector;
 import io.appform.ranger.drove.config.DroveUpstreamConfig;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import lombok.val;
-import org.apache.hc.client5.http.config.ConnectionConfig;
-import org.apache.hc.client5.http.config.RequestConfig;
-import org.apache.hc.client5.http.impl.classic.CloseableHttpClient;
-import org.apache.hc.client5.http.impl.classic.HttpClients;
-import org.apache.hc.client5.http.impl.io.PoolingHttpClientConnectionManagerBuilder;
-import org.apache.hc.client5.http.ssl.NoopHostnameVerifier;
-import org.apache.hc.client5.http.ssl.SSLConnectionSocketFactoryBuilder;
-import org.apache.hc.client5.http.ssl.TrustAllStrategy;
-import org.apache.hc.core5.ssl.SSLContextBuilder;
-import org.apache.hc.core5.util.TimeValue;
-import org.apache.hc.core5.util.Timeout;
-
-import java.util.List;
-import java.util.Objects;
 
 /**
  *
@@ -50,14 +31,6 @@ public class DroveNodeDataStoreConnector<T> implements NodeDataStoreConnector<T>
     protected final DroveUpstreamConfig config;
     protected final ObjectMapper mapper;
     protected final DroveClient droveClient;
-
-    public DroveNodeDataStoreConnector(
-            final DroveUpstreamConfig config,
-            final ObjectMapper mapper) {
-        this(config,
-             mapper,
-             buildDroveClient(config));
-    }
 
     public DroveNodeDataStoreConnector(
             final DroveUpstreamConfig config,
@@ -93,59 +66,4 @@ public class DroveNodeDataStoreConnector<T> implements NodeDataStoreConnector<T>
         return droveClient.leader().isPresent();
     }
 
-    @SneakyThrows
-    private static CloseableHttpClient createHttpClient(final DroveUpstreamConfig config) {
-        val connectionTimeout
-                = Objects.requireNonNullElse(config.getConnectionTimeout(),
-                                             DroveUpstreamConfig.DEFAULT_CONNECTION_TIMEOUT)
-                .toJavaDuration();
-        val cmBuilder = PoolingHttpClientConnectionManagerBuilder.create();
-        if (config.isInsecure()) {
-            log.debug("Creating insecure http client");
-            cmBuilder.setSSLSocketFactory(SSLConnectionSocketFactoryBuilder.create()
-                                                  .setSslContext(
-                                                          SSLContextBuilder.create()
-                                                                  .loadTrustMaterial(TrustAllStrategy.INSTANCE)
-                                                                  .build())
-                                                  .setHostnameVerifier(NoopHostnameVerifier.INSTANCE)
-                                                  .build());
-        }
-        val connectionManager = cmBuilder.build();
-        connectionManager.setDefaultMaxPerRoute(Integer.MAX_VALUE);
-        connectionManager.setMaxTotal(Integer.MAX_VALUE);
-        connectionManager.setDefaultConnectionConfig(ConnectionConfig.custom()
-                                                             .setConnectTimeout(Timeout.of(connectionTimeout))
-                                                             .setSocketTimeout(Timeout.of(connectionTimeout))
-                                                             .setValidateAfterInactivity(TimeValue.ofSeconds(10))
-                                                             .setTimeToLive(TimeValue.ofHours(1))
-                                                             .build());
-        val rc = RequestConfig.custom()
-                .setConnectionRequestTimeout(Timeout.of(connectionTimeout))
-                .setResponseTimeout(Timeout.of(Objects.requireNonNullElse(config.getOperationTimeout(),
-                                                                          DroveUpstreamConfig.DEFAULT_OPERATION_TIMEOUT)
-                                                       .toJavaDuration()))
-                .build();
-        return HttpClients.custom()
-                .disableRedirectHandling()
-                .setConnectionManager(connectionManager)
-                .setDefaultRequestConfig(rc)
-                .build();
-    }
-
-    private static DroveClient buildDroveClient(DroveUpstreamConfig config) {
-        val droveConfig = new DroveClientConfig(config.getEndpoints(),
-                                                Objects.requireNonNullElse(config.getCheckInterval(),
-                                                                           DroveUpstreamConfig.DEFAULT_CHECK_INTERVAL)
-                                                        .toJavaDuration(),
-                                                Objects.requireNonNullElse(config.getConnectionTimeout(),
-                                                                           DroveUpstreamConfig.DEFAULT_CONNECTION_TIMEOUT)
-                                                        .toJavaDuration(),
-                                                Objects.requireNonNullElse(config.getOperationTimeout(),
-                                                                           DroveUpstreamConfig.DEFAULT_OPERATION_TIMEOUT)
-                                                        .toJavaDuration());
-        return new DroveClient(droveConfig,
-                               List.of(new BasicAuthDecorator(config.getUsername(), config.getPassword()),
-                                       new AuthHeaderDecorator(config.getAuthHeader())),
-                               new DroveHttpComponentsTransport(droveConfig, createHttpClient(config)));
-    }
 }
