+++ ranger-drove/src/main/java/io/appform/ranger/drove/config/DroveUpstreamConfig.java
@@ -33,6 +33,7 @@ import java.util.List;
 public class DroveUpstreamConfig {
     public static final String DEFAULT_ENVIRONMENT_TAG_NAME = "environment";
     public static final String DEFAULT_REGION_TAG_NAME = "region";
+    public static final String DEFAULT_SKIP_TAG_NAME = "hidden";
     public static final Duration DEFAULT_CHECK_INTERVAL = Duration.seconds(10);
     public static final Duration DEFAULT_CONNECTION_TIMEOUT = Duration.seconds(5);
     public static final Duration DEFAULT_OPERATION_TIMEOUT = Duration.seconds(5);
@@ -54,6 +55,7 @@ public class DroveUpstreamConfig {
 
     String environmentTagName;
     String regionTagName;
+    String skipTagName;
 
     String defaultEnvironment;
     String defaultRegion;
